@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
@@ -19,16 +19,16 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.0" />
-    <PackageReference Include="Piranha" Version="10.4.0" />
-    <PackageReference Include="Piranha.AspNetCore" Version="10.4.0" />
-    <PackageReference Include="Piranha.AspNetCore.Identity" Version="10.4.0" />
-    <PackageReference Include="Piranha.AspNetCore.Identity.SQLite" Version="10.4.0" />
-    <PackageReference Include="Piranha.AttributeBuilder" Version="10.4.0" />
-    <PackageReference Include="Piranha.Data.EF" Version="10.4.0" />
-    <PackageReference Include="Piranha.Data.EF.SQLite" Version="10.4.0" />
-    <PackageReference Include="Piranha.Local.FileStorage" Version="10.4.0" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.1" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.1" />
+    <PackageReference Include="Piranha" Version="11.1.0" />
+    <PackageReference Include="Piranha.AspNetCore" Version="11.1.0" />
+    <PackageReference Include="Piranha.AspNetCore.Identity" Version="11.1.0" />
+    <PackageReference Include="Piranha.AspNetCore.Identity.SQLite" Version="11.1.0" />
+    <PackageReference Include="Piranha.AttributeBuilder" Version="11.1.0" />
+    <PackageReference Include="Piranha.Data.EF" Version="11.1.0" />
+    <PackageReference Include="Piranha.Data.EF.SQLite" Version="11.1.0" />
+    <PackageReference Include="Piranha.Local.FileStorage" Version="11.1.0" />
   </ItemGroup>
 
   <ItemGroup>
